Crate sdio_host

SD Card Registers

Register representations can be created from an array of little endian words. Note that the SDMMC protocol transfers the registers in big endian byte order.

let scr: SCR = [0, 1].into();

Reference documents:

PLSS_v7_10: Physical Layer Specification Simplified Specification Version 7.10. March 25, 2020. (C) SD Card Association

Modules

cmd

Structs

CIC

Card interface condition (R7)

CID

Card Identification Register (CID)

CSD

Card Specific Data (CSD)

CardStatus

Card Status (R1)

Cmd

Host to Card commands

OCR

Operation Conditions Register (OCR)

RCA

Relative Card Address (RCA)

SCR

SD CARD Configuration Register (SCR)

SDStatus

SD Status

Enums

BlockSize

BusWidth

The number of data lines in use on the SDMMC bus

CardCapacity

Types of SD Card

CurrentConsumption

CurrentState

CURRENT_STATE enum. Used for R1 response in command queue mode.

SDSpecVersion